test: move files in correct package
