+++ src/test/kotlin/io/github/filippovissani/dfrp/SemanticsSpec.kt
@@ -12,12 +12,12 @@ import io.github.filippovissani.dfrp.core.Semantics.neighbor
 import io.github.filippovissani.dfrp.core.Semantics.selfID
 import io.github.filippovissani.dfrp.core.Semantics.sense
 import io.github.filippovissani.dfrp.core.Then
-import io.github.filippovissani.dfrp.core.extensions.DefaultConfiguration.ELSE_VALUE
-import io.github.filippovissani.dfrp.core.extensions.DefaultConfiguration.LOCAL_SENSOR
-import io.github.filippovissani.dfrp.core.extensions.DefaultConfiguration.LOCAL_SENSOR_VALUE
-import io.github.filippovissani.dfrp.core.extensions.DefaultConfiguration.N_DEVICES
-import io.github.filippovissani.dfrp.core.extensions.DefaultConfiguration.THEN_VALUE
-import io.github.filippovissani.dfrp.core.extensions.Simulator.runSimulation
+import io.github.filippovissani.dfrp.DefaultConfiguration.ELSE_VALUE
+import io.github.filippovissani.dfrp.DefaultConfiguration.LOCAL_SENSOR
+import io.github.filippovissani.dfrp.DefaultConfiguration.LOCAL_SENSOR_VALUE
+import io.github.filippovissani.dfrp.DefaultConfiguration.N_DEVICES
+import io.github.filippovissani.dfrp.DefaultConfiguration.THEN_VALUE
+import io.github.filippovissani.dfrp.Simulator.runSimulation
 import io.github.filippovissani.dfrp.core.extensions.combine
 import io.github.filippovissani.dfrp.core.extensions.map
 import io.kotest.common.runBlocking
